feat: app release printer native
