+++ packages/desktop/electron-builder.yml
@@ -1,6 +1,7 @@
 ---
 appId: com.kalzate
 productName: Kalzate
+electronVersion : 2.0.8
 copyright: Copyright © 2020 ${author}
 win:
     target:
