+++ packages/desktop/public/package.json
@@ -12,5 +12,8 @@
   "license": "MIT",
   "bugs": {
     "url": "https://github.com/kalzate/kalzate/issues"
+  },
+  "dependencies": {
+    "printer": "^0.4.0"
   }
 }
