+++ packages/desktop/src/printer/template.html
@@ -43,9 +43,9 @@
 </head>
 <body>
 <section id="main"></section>
-<!-- <script src="../libs/jquery.min.js"></script>
+<script src="../libs/jquery.min.js"></script>
 <script src="../libs/jsbarcode.min.js"></script>
-<script src="../libs/qrcode.min.js"></script> -->
-<script src="./view.js"></script>
+<script src="../libs/qrcode.min.js"></script>
+<script src="./web-window-template.js"></script>
 </body>
 </html>
